refactor: reconfigure the storage engine (#1646)
* dev


* dev


* dev


* dev


* dev


* dev


* dev


---------

+++ src/common/metadata-struct/src/storage/shard.rs
@@ -27,7 +27,7 @@ pub struct EngineShard {
     pub config: EngineShardConfig,
     pub engine_type: EngineType,
     pub replica_num: u32,
-    pub create_time: u128,
+    pub create_time: u64,
 }
 
 impl EngineShard {
